requirements.txt
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0
transformers>=4.10.0
torch>=1.9.0
umap-learn>=0.5.2
tqdm>=4.62.0 